refactor: creat module view-mode
